rag_service/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML & Embeddings
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2

# Vector Database
faiss-cpu==1.7.4
numpy==1.24.3

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.3
PyPDF2==3.0.1

# Whisper STT (fallback for non-Indian languages)
openai-whisper==20231117

# IndicConformer for Indian language STT (primary - BhasaAnuvaad trained, 600M params, 22 languages)
# Uses transformers package above with AI4Bharat models
# Model: ai4bharat/indic-conformer-600m-multilingual

# Language Detection
langdetect==1.0.9

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
yfinance==0.2.28
pandas==2.1.4

# Audio Processing (for Whisper and IndicConformer)
ffmpeg-python==0.2.0
librosa==0.10.1
torchaudio==2.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Logging
loguru==0.7.2

# Optional: Advanced text processing
nltk==3.8.1
spacy==3.7.2
